add multiple activities and icons.

+++ src/com/appfirst/datatypes/BaseResourceData.java
@@ -24,13 +24,13 @@ import com.appfirst.types.BaseObject;
  * 
  */
 public class BaseResourceData extends BaseObject{
-	public int getTime() {
+	public long getTime() {
 		return time;
 	}
 
-	public void setTime(int time) {
+	public void setTime(long time) {
 		this.time = time;
 	}
 
-	private int time;
+	private long time;
 }
